+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+## master (unreleased)
+
+### `@self.id/core` v0.2.0
+
+Added support for `TileLoader` from `@glazed/tile-loader`.
+
+### `@self.id/framework` v0.1.0
+
+Initial release.
+
+### `@self.id/multiauth` v0.1.0
+
+Initial release.
+
+### `@self.id/react` v0.1.0
+
+Initial release.
+
+### `@self.id/ui` v0.1.0
+
+Initial release.
+
+### `@self.id/web` v0.2.0
+
+- Added support for `TileLoader` from `@glazed/tile-loader`.
+- Changed default behavior of attaching the DID instance to the Ceramic client once authenticated.
+
 ## 2020-09-15
 
-Release v0.1.0 of `@self.id/core`, `@self.id/web`, `@self.id/3box-legacy` and `@self.id/image-utils`.
+### `@self.id/3box-legacy` v0.1.0
+
+Initial release.
+
+### `@self.id/core` v0.1.0
+
+Initial release.
+
+### `@self.id/image-utils` v0.1.0
+
+Initial release.
+
+### `@self.id/web` v0.1.0
+
+Initial release.